Add files via upload
Morse Code is a method of encoding text using sequences of dots (.) and dashes (-) to represent letters, numbers, and symbols. It was originally used in telegraphy and is still used in aviation, emergency signaling, and amateur radio. For example:
A/a in Morse : .-
B/b in Morse : -...
Both lower and upper cases are treated the same in morse code.

What is 'Morseng' ? 
'Morseng' is a simple and efficient application that helps users to - 
1) convert english to morse code; type a message, tap Translate to get the morse equivalent instantly.
2) convert morse code to english; type a morse sequence and the app decodes this into plain text.
Morseng offers a clean, user-friendly interface. There’s no delay in processing—messages are translated and decoded instantly.

Whether you're learning Morse, sending secret messages, or just having fun, Morseng makes it seamless!